pyproject.toml
[project]
name = "energy-mlops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "entsoe-py>=0.8.0",
    "evidently>=0.7.21",
    "fastapi>=0.136.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "mlflow>=3.11.1",
    "polars>=1.40.0",
    "psycopg2-binary>=2.9.12",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.45.0",
]
